requirements.txt
langchain
python-dotenv
langchain-community
langchain_core
google-generativeai